refactor: extract shared Bridge Protocol types to @obsiscripta/shared
- Create packages/shared with common type definitions
- Move Bridge Protocol types (MCPContent, HealthResponse, Tool, etc.)
  to shared package for single source of truth
- Update obsidian-plugin to re-export from @obsiscripta/shared
- Update stdio-bridge to re-export from @obsiscripta/shared
- Add shared to pnpm-workspace.yaml
