+++ packages/obsidian-plugin/package.json
@@ -25,6 +25,7 @@
 		"typescript-eslint": "8.35.1"
 	},
 	"dependencies": {
+		"@obsiscripta/shared": "workspace:*",
 		"obsidian": "latest",
 		"sucrase": "^3.35.1"
 	}
