+++ packages/shared/package.json
@@ -1,0 +1,22 @@
+{
+	"name": "@obsiscripta/shared",
+	"version": "1.0.0",
+	"description": "Shared type definitions for ObsiScripta",
+	"type": "module",
+	"main": "dist/index.js",
+	"types": "dist/index.d.ts",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.js"
+		}
+	},
+	"scripts": {
+		"build": "tsc",
+		"dev": "tsc --watch"
+	},
+	"license": "0-BSD",
+	"devDependencies": {
+		"typescript": "^5.8.3"
+	}
+}
