+++ packages/stdio-bridge/package.json
@@ -23,7 +23,8 @@
 		"typescript": "^5.8.3"
 	},
 	"dependencies": {
-		"@modelcontextprotocol/sdk": "^1.0.0"
+		"@modelcontextprotocol/sdk": "^1.0.0",
+		"@obsiscripta/shared": "workspace:*"
 	},
 	"engines": {
 		"node": ">=16.0.0"
