+++ packages/stdio-bridge/src/types.ts
@@ -1,95 +1,27 @@
 /**
  * Type definitions for stdio-bridge
- * Bridge Protocol types and stdio-bridge specific types
+ * Re-exports Bridge Protocol types from @obsiscripta/shared
+ * and defines stdio-bridge specific types
  */
 
 // =============================================================================
-// MCPContent Type
+// Re-export Bridge Protocol Types from shared package
 // =============================================================================
 
-/**
- * MCP Content type - represents content in MCP format
- *
- * IMPORTANT: The `content` field relays the MCP format as-is (proxy behavior).
- * - No transformation or extension is performed
- * - This type is used to pass through MCP responses without modification
- *
- * Type: { type: "text" | "image" | ..., text?: string, data?: string, ... }
- */
-export type MCPContent = {
-	type: "text" | "image";
-	text?: string;
-	data?: string;
-	[key: string]: unknown;
-};
+export type {
+	MCPContent,
+	HealthResponse,
+	Tool,
+	ToolListResponse,
+	ToolCallRequest,
+	ToolCallSuccessResponse,
+	ToolCallErrorResponse,
+	ToolCallResponse,
+	ErrorResponse,
+} from "@obsiscripta/shared";
 
-// =============================================================================
-// Bridge Protocol Types (from plugin's bridge-types.ts)
-// =============================================================================
-
-/**
- * Health check response from the plugin
- */
-export interface HealthResponse {
-	status: "ok";
-	version: string;
-	protocolVersion: string;
-}
-
-/**
- * Tool definition exposed by the plugin
- */
-export interface Tool {
-	name: string;
-	description: string;
-	inputSchema: Record<string, unknown>;
-}
-
-/**
- * Response from GET /bridge/v1/tools
- */
-export interface ToolListResponse {
-	tools: Tool[];
-	hash: string;
-}
-
-/**
- * Request body for POST /bridge/v1/tools/{name}/call
- */
-export interface ToolCallRequest {
-	arguments: Record<string, unknown>;
-}
-
-/**
- * Successful tool call response
- */
-export interface ToolCallSuccessResponse {
-	success: true;
-	content: MCPContent[];
-}
-
-/**
- * Failed tool call response (tool execution error)
- */
-export interface ToolCallErrorResponse {
-	success: false;
-	content: MCPContent[];
-	isError: true;
-}
-
-/**
- * Tool call response - either success or error
- */
-export type ToolCallResponse = ToolCallSuccessResponse | ToolCallErrorResponse;
-
-/**
- * Error response for HTTP errors (404, 400, 500, etc.)
- */
-export interface ErrorResponse {
-	error: string;
-	message: string;
-	details?: unknown;
-}
+// Import MCPContent for use in this file
+import type { MCPContent } from "@obsiscripta/shared";
 
 // =============================================================================
 // stdio-bridge Specific Types
