+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 packages:
+  - packages/shared
   - packages/obsidian-plugin
   - packages/stdio-bridge
